app/auth.py
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from sqlalchemy.orm import Session
from sqlalchemy import or_
from app.database import get_db
from app.config import settings
from app import models, schemas
from app.security import security_manager

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt

def get_user_by_email(db: Session, email: str):
    """
    Trouve un utilisateur par email, en utilisant le hash pour la recherche
    """
    email_hash = security_manager.hash_value(email)
    return db.query(models.User).filter(models.User.email_hash == email_hash).first()


async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    """
    Récupère l'utilisateur courant à partir du token JWT.
    
    La fonction décode le token, extrait l'email, puis cherche l'utilisateur
    correspondant dans la base de données en utilisant le hash de l'email.
    """
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        # Décodage du token JWT
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    
    # Recherche de l'utilisateur par hash d'email
    email_hash = security_manager.hash_value(email)
    user = db.query(models.User).filter(models.User.email_hash == email_hash).first()
    
    if user is None:
        raise credentials_exception
    
    # Création d'un dictionnaire contenant les informations déchiffrées
    return schemas.User(
        id=user.id,
        email=security_manager.decrypt_value(user.email_encrypted),
        username= security_manager.decrypt_value(user.username_encrypted),
        phone=security_manager.decrypt_value(user.phone_encrypted),
        is_active=user.is_active,
        is_botanist=user.is_botanist
    )

def authenticate_user(db: Session, email: str, password: str):
    """Authenticate a user with email and password, handling encrypted fields."""
    user = get_user_by_email(db, email)
    if not user:
        return False
    if not verify_password(password, user.hashed_password):
        return False
    return user