requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Mako==1.3.9
MarkupSafe==3.0.2
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-instrumentation-sqlalchemy==0.55b0
opentelemetry-instrumentation-psycopg2==0.55b0
opentelemetry-instrumentation-asyncpg==0.55b0
opentelemetry-instrumentation-requests==0.55b0
opentelemetry-instrumentation-httpx==0.55b0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-propagator-b3==1.34.0
opentelemetry-propagator-jaeger==1.34.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.22.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0
python-json-logger==2.0.7